refactor(create-gatsby): Decouple package functions
