--- packages/create-gatsby/src/cmses.json.d.ts
@@ -1,5 +1,0 @@
-import {PluginMap} from "."
-
-declare const cmses: PluginMap
-
-export default cmses
