--- packages/create-gatsby/src/features.json.d.ts
@@ -1,5 +1,0 @@
-import {PluginMap} from "."
-
-declare const features: PluginMap
-
-export default features
