+++ packages/create-gatsby/src/index.ts
@@ -1,111 +1,28 @@
 import Enquirer from "enquirer"
-import cmses from "./cmses.json"
-import styles from "./styles.json"
-import features from "./features.json"
+import cmses from "./questions/cmses.json"
+import styles from "./questions/styles.json"
+import features from "./questions/features.json"
 import { initStarter, getPackageManager, gitSetup } from "./init-starter"
 import { installPlugins } from "./install-plugins"
-import c from "ansi-colors"
+import colors from "ansi-colors"
 import path from "path"
-import fs from "fs"
 import { plugin } from "./components/plugin"
 import { makePluginConfigQuestions } from "./plugin-options-form"
 import { center, wrap } from "./components/utils"
 import { stripIndent } from "common-tags"
 import { trackCli } from "./tracking"
-import crypto from "crypto"
-import { reporter } from "./reporter"
-import { setSiteMetadata } from "./site-metadata"
-import { makeNpmSafe } from "./utils"
-
-const sha256 = (str: string): string =>
-  crypto.createHash(`sha256`).update(str).digest(`hex`)
-
-const md5 = (str: string): string =>
-  crypto.createHash(`md5`).update(str).digest(`hex`)
-
-/**
- * Hide string on windows (for emojis)
- */
-const w = (input: string): string => (process.platform === `win32` ? `` : input)
-
-// eslint-disable-next-line no-control-regex
-const INVALID_FILENAMES = /[<>:"/\\|?*\u0000-\u001F]/g
-const INVALID_WINDOWS = /^(con|prn|aux|nul|com\d|lpt\d)$/i
+import { reporter } from "./utils/reporter"
+import { setSiteMetadata } from "./utils/site-metadata"
+import { makeNpmSafe } from "./utils/make-npm-safe"
+import {
+  validateProjectName,
+  generateQuestions,
+} from "./utils/question-helpers"
+import { sha256, md5 } from "./utils/hash"
+import { maybeUseEmoji } from "./utils/emoji"
 
 const DEFAULT_STARTER = `https://github.com/gatsbyjs/gatsby-starter-minimal.git`
 
-const makeChoices = (
-  options: Record<string, { message: string; dependencies?: Array<string> }>,
-  multi = false
-): Array<{ message: string; name: string; disabled?: boolean }> => {
-  const entries = Object.entries(options).map(([name, message]) => {
-    return { name, message: message.message }
-  })
-
-  if (multi) {
-    return entries
-  }
-  const none = { name: `none`, message: `No (or I'll add it later)` }
-  const divider = { name: `–`, role: `separator`, message: `–` }
-
-  return [none, divider, ...entries]
-}
-
-export const validateProjectName = async (
-  value: string
-): Promise<string | boolean> => {
-  if (!value) {
-    return `You have not provided a directory name for your site. Please do so when running with the 'y' flag.`
-  }
-  value = value.trim()
-  if (INVALID_FILENAMES.test(value)) {
-    return `The destination "${value}" is not a valid filename. Please try again, avoiding special characters.`
-  }
-  if (process.platform === `win32` && INVALID_WINDOWS.test(value)) {
-    return `The destination "${value}" is not a valid Windows filename. Please try another name`
-  }
-  if (fs.existsSync(path.resolve(value))) {
-    return `The destination "${value}" already exists. Please choose a different name`
-  }
-  return true
-}
-
-// The enquirer types are not accurate
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const questions = (initialFolderName: string, skip: boolean): any => [
-  {
-    type: `textinput`,
-    name: `project`,
-    message: `What would you like to name the folder where your site will be created?`,
-    hint: path.basename(process.cwd()),
-    separator: `/`,
-    initial: initialFolderName,
-    format: (value: string): string => c.cyan(value),
-    validate: validateProjectName,
-    skip,
-  },
-  {
-    type: `selectinput`,
-    name: `cms`,
-    message: `Will you be using a CMS?`,
-    hint: `(Single choice) Arrow keys to move, enter to confirm`,
-    choices: makeChoices(cmses),
-  },
-  {
-    type: `selectinput`,
-    name: `styling`,
-    message: `Would you like to install a styling system?`,
-    hint: `(Single choice) Arrow keys to move, enter to confirm`,
-    choices: makeChoices(styles),
-  },
-  {
-    type: `multiselectinput`,
-    name: `features`,
-    message: `Would you like to install additional features with other plugins?`,
-    hint: `(Multiple choice) Use arrow keys to move, spacebar to select, and confirm with an enter on "Done"`,
-    choices: makeChoices(features, true),
-  },
-]
 interface IAnswers {
   name: string
   project: string
@@ -141,10 +58,8 @@ export type PluginMap = Record<string, IPluginEntry>
 
 export type PluginConfigMap = Record<string, Record<string, unknown>>
 
-const removeKey = (plugin: string): string => plugin.split(`:`)[0]
-
 export async function run(): Promise<void> {
-  const [flag, siteDirectory] = process.argv.slice(2)
+  const [flag, siteDirectory] = process.argv.slice(2) // TODO - Refactor this to not be positional in upcoming TS PR since it's related
 
   let yesFlag = false
   if (flag === `-y`) {
@@ -155,13 +70,13 @@ export async function run(): Promise<void> {
 
   const { version } = require(`../package.json`)
 
-  reporter.info(c.grey(`create-gatsby version ${version}`))
+  reporter.info(colors.grey(`create-gatsby version ${version}`))
 
   reporter.info(
     `
 
 
-${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
+${center(colors.blueBright.bold.underline(`Welcome to Gatsby!`))}
 
 
 `
@@ -170,9 +85,9 @@ ${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
   if (!yesFlag) {
     reporter.info(
       wrap(
-        `This command will generate a new Gatsby site for you in ${c.bold(
+        `This command will generate a new Gatsby site for you in ${colors.bold(
           process.cwd()
-        )} with the setup you select. ${c.white.bold(
+        )} with the setup you select. ${colors.white.bold(
           `Let's answer some questions:\n\n`
         )}`,
         process.stdout.columns
@@ -186,16 +101,19 @@ ${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
 
   let data
   let siteName
+
   if (!yesFlag) {
     ;({ name: siteName } = await enquirer.prompt({
       type: `textinput`,
       name: `name`,
       message: `What would you like to call your site?`,
       initial: `My Gatsby Site`,
-      format: (value: string): string => c.cyan(value),
+      format: (value: string): string => colors.cyan(value),
     } as any))
 
-    data = await enquirer.prompt(questions(makeNpmSafe(siteName), yesFlag))
+    data = await enquirer.prompt(
+      generateQuestions(makeNpmSafe(siteName), yesFlag)
+    )
   } else {
     const warn = await validateProjectName(siteDirectory)
     if (typeof warn === `string`) {
@@ -204,7 +122,7 @@ ${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
     }
     siteName = siteDirectory
     data = await enquirer.prompt(
-      questions(makeNpmSafe(siteDirectory), yesFlag)[0]
+      generateQuestions(makeNpmSafe(siteDirectory), yesFlag)[0]
     )
   }
 
@@ -228,9 +146,9 @@ ${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
   })
 
   const messages: Array<string> = [
-    `${w(`🛠  `)}Create a new Gatsby site in the folder ${c.magenta(
-      data.project
-    )}`,
+    `${maybeUseEmoji(
+      `🛠  `
+    )}Create a new Gatsby site in the folder ${colors.magenta(data.project)}`,
   ]
 
   const plugins: Array<string> = []
@@ -239,7 +157,9 @@ ${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
 
   if (data.cms && data.cms !== `none`) {
     messages.push(
-      `${w(`📚 `)}Install and configure the plugin for ${c.magenta(
+      `${maybeUseEmoji(
+        `📚 `
+      )}Install and configure the plugin for ${colors.magenta(
         cmses[data.cms].message
       )}`
     )
@@ -255,7 +175,7 @@ ${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
 
   if (data.styling && data.styling !== `none`) {
     messages.push(
-      `${w(`🎨 `)}Get you set up to use ${c.magenta(
+      `${maybeUseEmoji(`🎨 `)}Get you set up to use ${colors.magenta(
         styles[data.styling].message
       )} for styling your site`
     )
@@ -272,8 +192,8 @@ ${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
 
   if (data.features?.length) {
     messages.push(
-      `${w(`🔌 `)}Install ${data.features
-        ?.map((feat: string) => c.magenta(feat))
+      `${maybeUseEmoji(`🔌 `)}Install ${data.features
+        ?.map((feat: string) => colors.magenta(feat))
         .join(`, `)}`
     )
     plugins.push(...data.features)
@@ -317,7 +237,7 @@ ${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
   if (!yesFlag) {
     reporter.info(`
 
-${c.bold(`Thanks! Here's what we'll now do:`)}
+${colors.bold(`Thanks! Here's what we'll now do:`)}
 
     ${messages.join(`\n    `)}
   `)
@@ -327,7 +247,7 @@ ${c.bold(`Thanks! Here's what we'll now do:`)}
       name: `confirm`,
       initial: `Yes`,
       message: `Shall we do this?`,
-      format: value => (value ? c.greenBright(`Yes`) : c.red(`No`)),
+      format: value => (value ? colors.greenBright(`Yes`) : colors.red(`No`)),
     })
 
     if (!confirm) {
@@ -341,16 +261,16 @@ ${c.bold(`Thanks! Here's what we'll now do:`)}
   await initStarter(
     DEFAULT_STARTER,
     data.project,
-    packages.map(removeKey),
+    packages.map((plugin: string) => plugin.split(`:`)[0]),
     siteName
   )
 
-  reporter.success(`Created site in ${c.green(data.project)}`)
+  reporter.success(`Created site in ${colors.green(data.project)}`)
 
   const fullPath = path.resolve(data.project)
 
   if (plugins.length) {
-    reporter.info(`${w(`🔌 `)}Setting-up plugins...`)
+    reporter.info(`${maybeUseEmoji(`🔌 `)}Setting-up plugins...`)
     await installPlugins(plugins, pluginConfig, fullPath, [])
   }
   await setSiteMetadata(fullPath, `title`, siteName)
@@ -362,22 +282,22 @@ ${c.bold(`Thanks! Here's what we'll now do:`)}
 
   reporter.info(
     stripIndent`
-    ${w(`🎉  `)}Your new Gatsby site ${c.bold(
+    ${maybeUseEmoji(`🎉  `)}Your new Gatsby site ${colors.bold(
       siteName
     )} has been successfully created
-    at ${c.bold(fullPath)}.
+    at ${colors.bold(fullPath)}.
     `
   )
   reporter.info(`Start by going to the directory with\n
-  ${c.magenta(`cd ${data.project}`)}
+  ${colors.magenta(`cd ${data.project}`)}
   `)
 
   reporter.info(`Start the local development server with\n
-  ${c.magenta(`${runCommand} develop`)}
+  ${colors.magenta(`${runCommand} develop`)}
   `)
 
   reporter.info(`See all commands at\n
-  ${c.blueBright(`https://www.gatsbyjs.com/docs/gatsby-cli/`)}
+  ${colors.blueBright(`https://www.gatsbyjs.com/docs/gatsby-cli/`)}
   `)
 
   const siteHash = md5(fullPath)
