+++ packages/create-gatsby/src/questions/cmses.json.d.ts
@@ -1,0 +1,5 @@
+import { PluginMap } from "../index"
+
+declare const cmses: PluginMap
+
+export default cmses
