+++ packages/create-gatsby/src/questions/features.json
