+++ packages/create-gatsby/src/questions/features.json.d.ts
@@ -1,0 +1,5 @@
+import { PluginMap } from "../index"
+
+declare const features: PluginMap
+
+export default features
