+++ packages/create-gatsby/src/questions/styles.json
