+++ packages/create-gatsby/src/questions/styles.json.d.ts
@@ -1,0 +1,5 @@
+import { PluginMap } from "../index"
+
+declare const styles: PluginMap
+
+export default styles
