--- packages/create-gatsby/src/styles.json.d.ts
@@ -1,5 +1,0 @@
-import {PluginMap} from "."
-
-declare const styles: PluginMap
-
-export default styles
