+++ packages/create-gatsby/src/utils/emoji.ts
@@ -1,0 +1,5 @@
+/**
+ * Hide string on windows (for emojis)
+ */
+export const maybeUseEmoji = (input: string): string =>
+  process.platform === `win32` ? `` : input
