+++ packages/create-gatsby/src/utils/get-config-store.ts
