+++ packages/create-gatsby/src/utils/is-truthy.ts
