+++ packages/create-gatsby/src/utils/site-metadata.ts
